+++ .gitignore
@@ -1,1 +1,3 @@
 notes/
+__pycache__
+__pycache__/
